separate gem loading benchmark
