+++ benchmarks/tracing_trace.rb
@@ -4,7 +4,6 @@ VALIDATE_BENCHMARK_MODE = ENV['VALIDATE_BENCHMARK'] == 'true'
 return unless __FILE__ == $PROGRAM_NAME || VALIDATE_BENCHMARK_MODE
 
 require 'benchmark/ips'
-require 'open3'
 require 'datadog'
 
 class TracingTraceBenchmark
@@ -177,39 +176,6 @@ class TracingTraceBenchmark
       end
     end
   end
-
-  def benchmark_gem_loading
-    # This benchmark needs to be run in a clean environment where datadog is not loaded yet
-    output, status = Open3.capture2e('bundle', 'exec', 'ruby', stdin_data: <<-RUBY)
-      raise "Datadog is already loaded" if defined?(::Datadog::Core)
-
-      lib = File.expand_path('../lib', '#{__dir__}')
-      $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-
-      VALIDATE_BENCHMARK_MODE = #{VALIDATE_BENCHMARK_MODE}
-      require 'benchmark/ips'
-
-      Benchmark.ips do |x|
-        # Gem loading is quite slower than the other microbenchmarks
-        benchmark_time = VALIDATE_BENCHMARK_MODE ? { time: 0.001, warmup: 0 } : { time: 60, warmup: 5 }
-        x.config(**benchmark_time)
-
-        x.report("Gem loading") do
-          pid = fork { require 'datadog' }
-
-          _, status = Process.wait2(pid)
-          raise unless status.success?
-        end
-
-        x.save! "#{__FILE__}-results.json" unless VALIDATE_BENCHMARK_MODE
-        x.compare!
-      end
-    RUBY
-
-    print output
-
-    raise "Benchmark failed with status #{status}: #{output}" unless status.success?
-  end
 end
 
 puts "Current pid is #{Process.pid}"
@@ -230,5 +196,4 @@ TracingTraceBenchmark.new.instance_exec do
   run_benchmark { benchmark_to_digest_continue }
   run_benchmark { benchmark_propagation_datadog }
   run_benchmark { benchmark_propagation_trace_context }
-  run_benchmark { benchmark_gem_loading }
 end
